fix(goto): read the browser version for client hints from the cache
Awaiting chromeVersionFromBrowser outside the action timeout could block the user agent override when browser.version() never settles. The client hints e2e tests move into test/index.js next to the other goto navigation tests.
